requirements.txt
beautifulsoup4==4.12.3
docker==7.1.0
fastapi==0.112.2
gunicorn==23.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.65
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.108
numpy==1.26.4
openpyxl==3.1.5
pgvector==0.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
safetensors==0.4.4
scipy==1.13.1
sentence-transformers==3.0.1
streamlit==1.38.0
tokenizers==0.19.1
torch==2.4.0
transformers==4.44.2
uvicorn==0.30.6
zipp==3.20.1
more-itertools==10.2.0
sentencepiece==0.2.0
opensearch-py==2.7.1